requirements.txt
psycopg2==2.9.1
redis==3.5.3
python-decouple==3.5
requests==2.26.0